fix: installed jest dependency, jest-config setup
